+++ lib/components/index.ts
@@ -5,11 +5,7 @@ export { GithubIcon } from "./Icons/GithubIcon";
 export { TwitterIcon } from "./Icons/TwitterIcon";
 export { GridContainer } from "./Layout/GridContainer";
 export { Section } from "./Layout/Section";
-export { HeaderDesktop } from "./Header/HeaderDesktop";
-export { HeaderMobile } from "./Header/HeaderMobile";
-export { NavItemDesktop } from "./Navigation/NavItemDesktop";
 export { NavMobile } from "./Navigation/NavMobile";
-export { NavItemMobile } from "./Navigation/NavItemMobile";
 export { ToggleNavButton } from "./Navigation/ToggleNavButton";
 export { ButtonPrimary } from "./Buttons/ButtonPrimary";
 export { LogoWithClaim } from "./Logos/LogoWithClaim";
